+++ package.json
@@ -9,6 +9,8 @@
     "nuka-carousel": "^8.0.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router": "^6.23.1",
+    "react-router-dom": "^6.23.1",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
