+++ src/components/NavBar/NavBar.jsx
@@ -1,12 +1,13 @@
 import React from "react";
 import './navbar.css';
+import {Link} from 'react-router-dom';
 
 function Navbar(){
     return (
         <nav className="navbar">
             <ul class="nav-list">
-                <a href="/">Home</a>
-                <a href="/About">About</a>
+                <Link to="/">Home</Link>
+                <Link to="/About">About</Link>
                 <a href="/Project">Project</a>
                 <a href="/Resume">Resume</a>
             </ul>
