+++ src/components/Pages/AboutPage/AboutPage.jsx
@@ -1,0 +1,9 @@
+function AboutPage(){
+    return(
+        <div>
+            About Page
+        </div>
+    )
+}
+
+export default AboutPage;
